Create the AnimeList component
